Corrected use of int vs uint64 for frame number in extract_frame.

+++ cmd/extract_frame/extract_frame.go
@@ -83,7 +83,7 @@ func main() {
 		}
 
 		// Try extracting a frame
-		img, err := mov.ExtractFrame(i)
+		img, err := mov.ExtractFrame(uint64(i))
 
 		if err != nil {
 			fmt.Printf("Error decoding frame: %s\n", err.Error())
